Title: 차이가 가장 작은 수, Time: -1ms, Memory: -1MB, Status: Compile Error - CodeTree

+++ 241212/차이가 가장 작은 수/the-number-with-the-smallest-difference.cpp
@@ -6,7 +6,7 @@ using namespace std;
 
 int n, m;
 vector<int> vec;
-int answer = INT_MAX;
+long answer = INT_MAX;
 
 int main()
 {
